requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.4.0
pytesseract==0.3.13
easyocr==1.7.2 
opencv-python-headless==4.12.0.88